third_party/blink/web_tests/platform/generic/fast/replaced/vertical-resize-100percent-element-expected.txt

Layout test for https://bugs.webkit.org/show_bug.cgi?id=43022. Checks to see if an image having a percentage height is resized when its containing iframe is vertically resized. This test only works in DumpRenderTree since it involves accessing the internal elements of an iframe.

Original frame height was 500
Original image height was 500
Vertically resizing the frame to be larger
New frame height is 600
New image height is 600
PASSED: Image size is correct
Vertically resizing the frame to be smaller
New frame height is 450
New image height is 450
PASSED: Image size is correct
